requirements.txt
spacy==2.1.3
neuralcoref==4.0
pytorch-pretrained-bert==0.6.1
tensorboard==1.13.1

